+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 * Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
 * Added optional storage of tests in notebook metadata for Otter Assign
+* Removed deprecated tool Otter Service
 
 **v2.2.2:**