+++ docs/changelog.md
@@ -3,7 +3,7 @@
 **Unreleased:**
 
 * Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
-* Added optional storage of tests in notebook metadata for Otter Assign
+* Added optional storage of tests in notebook metadata for Otter Assign (default-off until v3)
 * Removed deprecated tool Otter Service
 * Added printout of total score to autograder output