+++ docs/changelog.md
@@ -3,8 +3,9 @@
 **Unreleased:**
 
 * Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
-* Added optional storage of tests in notebook metadata for Otter Assign
+* Added optional storage of tests in notebook metadata for Otter Assign (default-off until v3)
 * Removed deprecated tool Otter Service
+* Added printout of total score to autograder output
 
 **v2.2.2:**